Ejercicio 16 completado y 17 a medias

+++ Tema1/Ejercicio14.java
@@ -13,7 +13,7 @@ public class Ejercicio14 {
     numRan = Math.ceil(numRan);
 
     // Imprimimos en terminal la variable numRan
-    System.out.println(numRan);
+    System.out.println((int)numRan);
   }
 
 }
